chore: update and replace breaking zora package
zora's author changed their release from 2.x to 3.x but also pulled it from NPM as a now beta version instead of the standard version that's been available for the past 10 months. Even using the latest @next version is failing due to broken code in zora itself. Replaced the entire library with **muggle**. This new library doesn't have the same level of asynchronisity as zora, but as we are only testing pipelines its not necessary and its assertion approach is much simpler.